# week10/dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"UserMissionAddResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "유저 미션 ID"
					},
					"userId": {
						"type": "number",
						"format": "double",
						"description": "유저 ID"
					},
					"missionId": {
						"type": "number",
						"format": "double",
						"description": "미션 ID"
					},
					"status": {
						"type": "string",
						"description": "미션 상태 (CHALLENGING / DONE)"
					}
				},
				"required": [
					"id",
					"userId",
					"missionId",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"StoreAddResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "가게 ID"
					},
					"name": {
						"type": "string",
						"description": "가게 이름"
					},
					"regionId": {
						"type": "number",
						"format": "double",
						"description": "지역 ID"
					}
				},
				"required": [
					"id",
					"name",
					"regionId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_StoreAddResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/StoreAddResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StoreAddApiResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/StoreAddResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StoreAddRequest": {
				"properties": {
					"name": {
						"type": "string",
						"description": "가게 이름",
						"example": "맛있는 한식당"
					},
					"foodCategoryId": {
						"type": "number",
						"format": "double",
						"description": "음식 카테고리 ID",
						"example": 1
					},
					"regionId": {
						"type": "number",
						"format": "double",
						"description": "지역 ID",
						"example": 1
					},
					"address": {
						"type": "string",
						"description": "가게 주소",
						"example": "서울시 강남구 역삼동 123-45"
					},
					"latitude": {
						"type": "number",
						"format": "double",
						"description": "위도",
						"example": 37.5012
					},
					"longitude": {
						"type": "number",
						"format": "double",
						"description": "경도",
						"example": 127.0396
					}
				},
				"required": [
					"name",
					"foodCategoryId",
					"regionId",
					"address",
					"latitude",
					"longitude"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewAddResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "리뷰 ID"
					},
					"storeId": {
						"type": "number",
						"format": "double",
						"description": "가게 ID"
					},
					"userId": {
						"type": "number",
						"format": "double",
						"description": "유저 ID"
					},
					"rating": {
						"type": "number",
						"format": "double",
						"description": "별점"
					},
					"content": {
						"type": "string",
						"nullable": true,
						"description": "리뷰 내용"
					}
				},
				"required": [
					"id",
					"storeId",
					"userId",
					"rating",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ReviewAddResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/ReviewAddResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewAddApiResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/ReviewAddResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewAddRequest": {
				"properties": {
					"rating": {
						"type": "number",
						"format": "double",
						"description": "별점 (1.0 ~ 5.0)",
						"example": 4.5
					},
					"content": {
						"type": "string",
						"description": "리뷰 내용 (선택)",
						"example": "정말 맛있어요!"
					}
				},
				"required": [
					"rating"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewListResponse": {
				"properties": {
					"reviews": {
						"items": {
							"$ref": "#/components/schemas/ReviewAddResponse"
						},
						"type": "array",
						"description": "리뷰 목록"
					}
				},
				"required": [
					"reviews"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ReviewListResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/ReviewListResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewListApiResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/ReviewListResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserMissionAddResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UserMissionAddResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserMissionAddApiResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UserMissionAddResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionAddResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "미션 ID"
					},
					"storeId": {
						"type": "number",
						"format": "double",
						"description": "가게 ID"
					},
					"content": {
						"type": "string",
						"description": "미션 내용"
					},
					"rewardPoint": {
						"type": "number",
						"format": "double",
						"description": "리워드 포인트"
					},
					"deadline": {
						"type": "number",
						"format": "double",
						"nullable": true,
						"description": "마감 기한"
					}
				},
				"required": [
					"id",
					"storeId",
					"content",
					"rewardPoint",
					"deadline"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_MissionAddResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/MissionAddResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionAddApiResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/MissionAddResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionAddRequest": {
				"properties": {
					"content": {
						"type": "string",
						"description": "미션 내용",
						"example": "1만원 이상 주문하기"
					},
					"rewardPoint": {
						"type": "number",
						"format": "double",
						"description": "미션 달성 시 지급되는 포인트",
						"example": 500
					},
					"deadline": {
						"type": "number",
						"format": "double",
						"description": "미션 마감 기한 (선택)",
						"example": 30
					}
				},
				"required": [
					"content",
					"rewardPoint"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionListResponse": {
				"properties": {
					"missions": {
						"items": {
							"$ref": "#/components/schemas/MissionAddResponse"
						},
						"type": "array",
						"description": "미션 목록"
					}
				},
				"required": [
					"missions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_MissionListResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/MissionListResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionListApiResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/MissionListResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "playground-umc-10th-nodejs",
		"version": "1.0.0",
		"description": "UMC 테스트 프로젝트",
		"contact": {}
	},
	"paths": {
...
		"/stores": {
			"post": {
				"operationId": "HandleAddStore",
				"responses": {
					"200": {
						"description": "가게 추가 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_StoreAddResponse_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"statusCode": 200,
											"message": "가게 추가 성공",
											"data": {
												"id": 1,
												"name": "맛있는 한식당",
												"regionId": 1
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "로그인이 필요합니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StoreAddApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 401,
											"message": "로그인이 필요합니다.",
											"data": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "서버 내부 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StoreAddApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 500,
											"message": "서버 내부 오류",
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"description": "가게 추가 API",
				"summary": "특정 지역에 가게를 추가하는 엔드포인트입니다.",
				"tags": [
					"Stores"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/StoreAddRequest"
							}
						}
					}
				}
			}
		},
		"/stores/{storeId}/reviews": {
			"post": {
				"operationId": "HandleAddReview",
				"responses": {
					"200": {
						"description": "리뷰 추가 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ReviewAddResponse_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"statusCode": 200,
											"message": "리뷰 추가 성공",
											"data": {
												"id": 1,
												"storeId": 1,
												"userId": 1,
												"rating": 4.5,
												"content": "정말 맛있어요!"
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "로그인이 필요합니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ReviewAddApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 401,
											"message": "로그인이 필요합니다.",
											"data": null
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 가게 또는 사용자입니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ReviewAddApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 404,
											"message": "존재하지 않는 가게 또는 사용자입니다.",
											"data": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "서버 내부 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ReviewAddApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 500,
											"message": "서버 내부 오류",
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"description": "리뷰 추가 API",
				"summary": "특정 가게에 리뷰를 추가하는 엔드포인트입니다.",
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReviewAddRequest"
							}
						}
					}
				}
			}
		},
		"/users/reviews": {
			"get": {
				"operationId": "HandleGetUserReviews",
				"responses": {
					"200": {
						"description": "리뷰 목록 조회 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ReviewListResponse_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"statusCode": 200,
											"message": "리뷰 목록 조회 성공",
											"data": {
												"reviews": [
													{
														"id": 1,
														"storeId": 1,
														"userId": 1,
														"rating": 4.5,
														"content": "정말 맛있어요!"
													}
												]
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "로그인이 필요합니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ReviewListApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 401,
											"message": "로그인이 필요합니다.",
											"data": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "서버 내부 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ReviewListApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 500,
											"message": "서버 내부 오류",
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"description": "내 리뷰 목록 조회 API",
				"summary": "로그인한 사용자가 작성한 리뷰 목록을 조회하는 엔드포인트입니다.",
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": []
			}
		},
		"/missions/{missionId}/challenge": {
			"post": {
				"operationId": "HandleAddUserMission",
				"responses": {
					"200": {
						"description": "미션 도전 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserMissionAddResponse_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"statusCode": 200,
											"message": "미션 도전 성공",
											"data": {
												"id": 1,
												"userId": 1,
												"missionId": 1,
												"status": "CHALLENGING"
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "로그인이 필요합니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserMissionAddApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 401,
											"message": "로그인이 필요합니다.",
											"data": null
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 사용자 또는 미션입니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserMissionAddApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 404,
											"message": "존재하지 않는 사용자 또는 미션입니다.",
											"data": null
										}
									}
								}
							}
						}
					},
					"409": {
						"description": "이미 도전 중인 미션입니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserMissionAddApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 409,
											"message": "이미 도전 중인 미션입니다.",
											"data": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "서버 내부 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserMissionAddApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 500,
											"message": "서버 내부 오류",
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"description": "미션 도전 API",
				"summary": "사용자가 특정 미션에 도전하는 엔드포인트입니다.",
				"tags": [
					"Missions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "missionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/missions/stores/{storeId}": {
			"post": {
				"operationId": "HandleAddMission",
				"responses": {
					"200": {
						"description": "미션 추가 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_MissionAddResponse_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"statusCode": 200,
											"message": "미션 추가 성공",
											"data": {
												"id": 1,
												"storeId": 1,
												"content": "1만원 이상 주문하기",
												"rewardPoint": 500,
												"deadline": 30
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "로그인이 필요합니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MissionAddApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 401,
											"message": "로그인이 필요합니다.",
											"data": null
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 가게입니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MissionAddApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 404,
											"message": "존재하지 않는 가게입니다.",
											"data": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "서버 내부 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MissionAddApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 500,
											"message": "서버 내부 오류",
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"description": "미션 추가 API",
				"summary": "특정 가게에 미션을 추가하는 엔드포인트입니다.",
				"tags": [
					"Missions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MissionAddRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "HandleGetStoreMissions",
				"responses": {
					"200": {
						"description": "미션 목록 조회 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_MissionListResponse_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"statusCode": 200,
											"message": "미션 목록 조회 성공",
											"data": {
												"missions": [
													{
														"id": 1,
														"storeId": 1,
														"content": "1만원 이상 주문하기",
														"rewardPoint": 500,
														"deadline": 30
													}
												]
											}
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 가게입니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MissionListApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 404,
											"message": "존재하지 않는 가게입니다.",
											"data": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "서버 내부 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MissionListApiResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": false,
											"statusCode": 500,
											"message": "서버 내부 오류",
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"description": "가게 미션 목록 조회 API",
				"summary": "특정 가게의 미션 목록을 조회하는 엔드포인트입니다.",
				"tags": [
					"Missions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "페이지네이션 커서",
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}